Add shooting functionality
There are some class changes in this commit. Weapon is now an abstract
class with an abstract method ShootFrom which takes a position and a
direction as arguments. This was done to achieve better abstraction
when developing new weapons.
Upon clicking the fire button the ShootFrom method from the current
weapon in PlayerWeapons is called with the world position of the
center of the screen and forward vector of the main camera.
The only implemented weapon so far is Pistol, which casts a ray
forward upon firing.