Implement Zombieman movement
Added the ZombiemanMovement script, which uses a NavMeshAgent component
to move the Zombieman to a given position. The script
RelativeVelocityCalculator is used to calculate the velocity of
a GameObject relative to a target Transform component and then
write the result into the properties of the Animator of the
GameObject. Using this script different walking animations are
played for the Zombieman, depending on the direction he is headed.